wdc/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
parameterized